[Tabs][Carousel] Verify / support a nested tabs scenario  * Removed the changes done for Json export. These changes will be included in a new branch.
