+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/sandbox/internal/models/CarouselImpl.java
@@ -21,11 +21,11 @@ import org.apache.sling.models.annotations.Model;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.adobe.cq.export.json.ContainerExporter;
+import com.adobe.cq.export.json.ComponentExporter;
 import com.adobe.cq.export.json.ExporterConstants;
 import com.adobe.cq.wcm.core.components.sandbox.models.Carousel;
 
-@Model(adaptables = SlingHttpServletRequest.class, adapters = {Carousel.class, ContainerExporter.class}, resourceType = CarouselImpl.RESOURCE_TYPE)
+@Model(adaptables = SlingHttpServletRequest.class, adapters = {Carousel.class, ComponentExporter.class}, resourceType = CarouselImpl.RESOURCE_TYPE)
 @Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME, extensions = ExporterConstants.SLING_MODEL_EXTENSION)
 public class CarouselImpl extends AbstractContainerImpl implements Carousel {
