+++ bundles/core/src/test/java/com/adobe/cq/wcm/core/components/sandbox/internal/models/TabsImplTest.java
@@ -26,13 +26,11 @@ import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
 
-import com.adobe.cq.export.json.SlingModelFilter;
 import com.adobe.cq.sightly.WCMBindings;
 import com.adobe.cq.wcm.core.components.Utils;
 import com.adobe.cq.wcm.core.components.context.CoreComponentTestContext;
 import com.adobe.cq.wcm.core.components.models.ListItem;
 import com.adobe.cq.wcm.core.components.sandbox.models.Tabs;
-import com.adobe.cq.wcm.core.components.testing.MockSlingModelFilter;
 import io.wcm.testing.mock.aem.junit.AemContext;
 
 import static org.junit.Assert.assertEquals;
@@ -53,7 +51,6 @@ public class TabsImplTest {
     @BeforeClass
     public static void init() {
         AEM_CONTEXT.load().json(TEST_BASE + CoreComponentTestContext.TEST_APPS_JSON, TEST_APPS_ROOT);
-        AEM_CONTEXT.registerService(SlingModelFilter.class, new MockSlingModelFilter());
     }
 
     @Test
